Consolidate ingest service tests
