+++ .github/workflows/ci.yml
@@ -15,6 +15,6 @@ jobs:
           pip install -r services/ingest/requirements.txt
           pip install pytest flake8
       - name: Lint
-        run: flake8 services tests
+        run: flake8 services
       - name: Pytest
         run: pytest -q
