+++ services/ingest/tests/test_api.py
@@ -9,3 +9,10 @@ def test_post_alerts():
     response = client.post("/alerts", json=payload)
     assert response.status_code == 200
     assert response.json() == {"status": "received"}
+
+
+def test_receive_alert():
+    payload = {"id": "1", "description": "test", "severity": 5}
+    response = client.post("/alerts", json=payload)
+    assert response.status_code == 200
+    assert response.json() == {"status": "received"}
